Move launching of UniversePackageBuilder back to universe_builder.py
